Arnold/app/requirements.txt
# for additional pywaggle install options, see: https://github.com/waggle-sensor/pywaggle#installation-guides
pywaggle[vision]==0.56.0
numpy==1.24.3
pytorch==2.0.1
pytorch-cuda==11.7
torchvision==0.15.2

